Automatic fencing configuration
Implement the "baremetal fencing export" TripleO command
that converts an instack.json file into a yaml file that
can be passed to the overcloud deploy command and that
preconfigures fencing on the nodes of the cluster.

Parsing the JSON file has been preferred over using the
Ironic API because by default ironic does not allow the
retrieval of passwords in cleartext.

Currently only the pxe_ssh and the pxe_ipmitool drivers
are supported because those are the only supported modules
within the tripleo puppet fencing modules.